refactored locations so I could recode the entire project
