+++ FtcRobotController/build.gradle
@@ -9,8 +9,8 @@ android {
 
     defaultConfig {
         minSdkVersion 24
+        targetSdk 33
         //noinspection ExpiredTargetSdkVersion
-        targetSdkVersion 28
         buildConfigField "String", "APP_BUILD_TIME", '"' + (new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ", Locale.ROOT).format(new Date())) + '"'
     }
