+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/src/v1/DriveController.java
@@ -1,7 +1,6 @@
-package org.firstinspires.ftc.teamcode.src;
+package org.firstinspires.ftc.teamcode.src.v1;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
 enum ModuleSide{LEFT, RIGHT}
