+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/src/v1/Vector2d.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.src;
+package org.firstinspires.ftc.teamcode.src.v1;
 
 //credit: this class is based on code from FRC 5818 (https://github.com/Team5818/DiffSwerve)
